requirements.txt
google-generativeai>=0.5.0
linkedin-api>=2.0.0
chromadb>=0.4.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
loguru>=0.7.0
tenacity>=8.2.0
python-dateutil>=2.8.0
pytest>=7.4.0
